Fix UI test ownership boundary
